Release v1.3.0 Dev into Master

Closes #29, #25, #24, #35, #33, #31, #30, #26, #23, #8, #36, #16, and #37

See merge request wavelab/wavelab!29

+++ wavelab/gui/baro_pressure_gui.py
@@ -1,21 +1,25 @@
 #!/usr/bin/env python3
 """
 Contains a GUI that interfaces with this package's netCDF-writing
-modules. Also contains some convenience methods for other GUIs and a
-general class, MessageDialog, for creating custom dialog boxes.
+modules. Also contains some convenience methods for other GUIs.
 """
 
 import tkinter as tk
 from tkinter import ttk
 from tkinter.filedialog import askopenfilename
+import pathlib
 import os
 from collections import OrderedDict
 from wavelab.processing.pressure_script import convert_to_netcdf
+from wavelab.utilities.utils import MessageDialog
 import json
 import re
 
-GLOBAL_HISTFILE = 'history.json'
-LOCAL_HISTFILE = 'history2_air.json'
+# Save history JSON files at a location: ex. C:\Users\username\WaveLab
+HISTFILEPATH = os.environ['USERPROFILE'] + r'\WaveLab'
+pathlib.Path(HISTFILEPATH).mkdir(parents=True, exist_ok=True) # Create WaveLab folder if it doesn't already exist
+GLOBAL_HISTFILE = os.path.join(HISTFILEPATH, 'history.json') 
+LOCAL_HISTFILE = os.path.join(HISTFILEPATH, 'history2_air.json')
 GLOBAL_FIELDS = OrderedDict([
     ('creator_name', ['Your full name:', '']),
     ('creator_email', ['Your email address:', '']),
@@ -23,7 +27,7 @@ GLOBAL_FIELDS = OrderedDict([
 LOCAL_FIELDS = OrderedDict([
     ('instrument_name', ['Instrument:', [
         'MS TruBlue 255', 'Onset Hobo U20', 'LevelTroll', 'RBRSolo',
-        'NOAA Station', 'Meso West', 'VanEssen'#, 'USGS Homebrew'
+        'NOAA Station', 'Meso West', 'VanEssen' # , 'Generic', 'USGS Homebrew'
         ], True]),
      ('stn_station_number', ['STN Site Id:', '']),
     ('stn_instrument_id', ['STN Instrument Id:', '']),
@@ -96,52 +100,55 @@ class BaroPressureGUI:
         message = ('Working, this may take a few minutes.')
         dialog = None
 
-        try:
-            dialog = MessageDialog(self.parent, message=message,
-                                   title='Processing...', buttons=0, wait=False)
-            globs = dict(zip(GLOBAL_FIELDS.keys(),
-                             self.global_form.export_entries()))
+        #  If no file is selected yet
+        if len(self.datafiles) == 0:
+            MessageDialog(self.parent, message='No file selected. Please select a file before running.',
+                                  title='No file selected!')
+        # Run if there is a file
+        else:
+            try:
+                dialog = MessageDialog(self.parent, message=message,
+                                    title='Processing...', buttons=0, wait=False)
+                globs = dict(zip(GLOBAL_FIELDS.keys(),
+                                self.global_form.export_entries()))
 
 
-            for fname, datafile in self.datafiles.items():
-                inputs = dict(zip(LOCAL_FIELDS.keys(), datafile.export_entries()))
-                inputs.update(globs)
-                if self.air_pressure == False:
-                    inputs['pressure_type'] = 'Sea Pressure'
-                else:
-                    inputs['pressure_type'] = 'Air Pressure'
-                inputs['in_filename'] = fname
-                inputs['out_filename'] = fname + '.nc'
+                for fname, datafile in self.datafiles.items():
+                    inputs = dict(zip(LOCAL_FIELDS.keys(), datafile.export_entries()))
+                    inputs.update(globs)
+                    if self.air_pressure == False:
+                        inputs['pressure_type'] = 'Sea Pressure'
+                    else:
+                        inputs['pressure_type'] = 'Air Pressure'
+                    inputs['in_filename'] = fname
+                    inputs['out_filename'] = fname + '.nc'
 
-                process_files = self.validate_entries(inputs)
+                    process_files = self.validate_entries(inputs)
 
-                if process_files == True:
-                    convert_to_netcdf(inputs)
-                    self.remove_file(fname)
+                    if process_files == True:
+                        bad_data, message = convert_to_netcdf(inputs)
+                        if bad_data:
+                            MessageDialog(self.parent, message=message,
+                                title='Data Issues!')
+                        else:
+                            self.remove_file(fname)
+                            dialog.destroy()
+                            MessageDialog(self.parent, message="Success! Files saved.",
+                                title='Success!')
+                    else:
+                        dialog.destroy()
+                        MessageDialog(self.parent, message= self.error_message,
+                                title='Error')
+
+                    self.error_message = ''
+
+            except:
+                if dialog is not None:
                     dialog.destroy()
-                    MessageDialog(self.parent, message="Success! Files saved.",
-                          title='Success!')
-                else:
-                    dialog.destroy()
-                    MessageDialog(self.parent, message= self.error_message,
-                              title='Error')
 
-                self.error_message = ''
+                MessageDialog(self.parent, message="Could not process files, please check file type.",
+                            title='Error')
 
-        except:
-            if dialog is not None:
-                dialog.destroy()
-
-            MessageDialog(self.parent, message="Could not process files, please check file type.",
-                        title='Error')
-            # import sys, traceback
-            # exc_type, exc_value, exc_traceback = sys.exc_info()
-            #
-            # message = traceback.format_exception(exc_type, exc_value,
-            #                               exc_traceback)
-            # MessageDialog(self.parent, message=message,
-            #               title='Error')
-            
     
     def validate_entries(self, inputs):
         '''Check if the GUI entries are filled out and in the proper format'''
@@ -279,40 +286,6 @@ class ButtonBar(tk.Frame):
         for i, props in enumerate(buttonlist):
             button = tk.Button(self, text=props[0], command=props[1], width=12)
             button.grid(row=0, column=i, sticky=('W', 'E'))
-
-
-class MessageDialog(tk.Toplevel):
-    """ A template for nice dialog boxes. """
-
-    def __init__(self, parent, message="", title="", buttons=1, wait=True):
-        tk.Toplevel.__init__(self, parent)
-        body = tk.Frame(self)
-        self.title(title)
-        self.boolean = None
-        self.parent = parent
-        self.transient(parent)
-        tk.Label(body, text=message).pack()
-        if buttons == 1:
-            b = tk.Button(body, text="OK", command=self.destroy)
-            b.pack(pady=5)
-        elif buttons == 2:
-            buttonframe = tk.Frame(body, padding="3 3 5 5")
-            def event(boolean):
-                self.boolean = boolean
-                self.destroy()
-            b1 = tk.Button(buttonframe, text='YES',
-                           command=lambda: event(True))
-            b1.grid(row=0, column=0)
-            b2 = tk.Button(buttonframe, text='NO',
-                           command=lambda: event(False))
-            b2.grid(row=0, column=1)
-            buttonframe.pack()
-        body.pack()
-        self.grab_set()
-        self.geometry("+%d+%d" % (parent.winfo_rootx()+50,
-                                  parent.winfo_rooty()+50))
-        if wait:
-            self.wait_window(self)
 
 
 if __name__ == '__main__':
